create manage component and change location of ManageMenu (old name: LeftMenu)
