+++ client_side/src/components/manage/manageIngredient/ManageIngredient.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ManageIngredient = () => {
+    return <div>Hello Ingredient</div>;
+};
+
+export default ManageIngredient;
