+++ client_side/src/components/manage/manageMenu/TemplateData.jsx
